Move `create_prepub` to the devdata module
This allows it to be more easily shared.
